backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
pytest==7.4.3
pytest-django==4.7.0

# Enhanced dependencies for user management and social features
django-allauth==0.57.0
celery==5.3.4
redis==5.0.1
pillow==10.1.0
django-storages==1.14.2
django-redis==5.4.0
gunicorn==21.2.0

# Monitoring and logging dependencies
psutil==5.9.6

# Testing and coverage dependencies
pytest-cov==4.1.0
coverage==7.3.2
factory-boy==3.3.0
faker==20.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
locust==2.17.0